Update CORTXContributionGuide.md
Do we have more than one code style guide? If so, do they have any commonalities? It would probably be good to have a blurb summarizing them if we can, with links to the various.  There's also this sentence in the link to the code style guide (" This document does NOT describe code documentation practices, they are explained elsewhere.") and if there are documentation processes, we should probably include them.

I think also we might want a section on "how to file a bug" or "how to suggest a feature or improvement" in addition to a section that explains the triage process (which I know is TBD, as are these other things.)