adds initial commit for spark integration for the devpost/cortx hackathon.
This commit includes a link to a video (youtube), installation documentation, a bit of background on Spark/Hadoop and ansible script for easy installation of the integration.

This commit was tested using the CORTX OVA image using VmWare Workstation for the CORTX provider, and an Ubuntu 20.04 image running under VmWare Workstation for the integration.

Signed-off-by: Stephen M <stephen@motts.id.au>